line progress on width

+++ Application/assets/shaders/line.vert
@@ -1,7 +1,7 @@
 #version 330 core
 
 layout (location = 0) in vec3 aPosition;
-layout (location = 1) in int aIndex;
+layout (location = 1) in float aIndex;
 
 uniform mat4 aTransform;
 uniform vec4 aColor;
@@ -13,7 +13,7 @@ layout (std140) uniform camera
 	float uWorldToScreen;
 };
 
-out int vIndex;
+out float vIndex;
 
 void main()
 {
